Compose transformations

Compute the composition of two or more transforms, the single transform that combines their effects in order.

composeTransforms(...)

Arguments

...
Affine or nonlinear transforms, possibly obtained from forward or reverse.

Value

The composed transform. If all arguments are affines then the result will also be an affine; otherwise it will be a deformation field.

Note

The source image for the composed transform is generally the source image from the first transform, and the target is the target image from the second transform. However, the target image attached to half transforms (as calculated by halfTransform) generally has a modified xform, compared to the original target. Therefore, composing a half transform with itself may not be exactly equivalent to the original.

See also

Author

Jon Clayden <code@clayden.org>