Save and load transform objects

These objects save a full transformation object, including source and target image metadata, to a self-contained RDS file, or load it back from such a file.

saveTransform(transform, fileName = NULL)

loadTransform(x)

Arguments

transform

A transform, possibly obtained from forward or reverse.

fileName

The file name to save to. If NULL, the serialised object is returned directly instead.

x

A file name to read from, or a serialised transform object.

Value

saveTransform returns a serialised transform object, if no filename is given; otherwise it is called for its side-effect of writing to file. loadTransform returns a deserialised transform object.

See also

Author

Jon Clayden <code@clayden.org>