Similarity measures between images

This function calculates a similarity measure between two images, after resampling one into the space of the other. The only supported measure is currently normalised mutual information, which is also used as a cost function by the registration algorithms.

similarity(source, target, targetMask = NULL, interpolation = 3L,
  threads = getOption("RNiftyReg.threads"))

Arguments

source

The source image, in any acceptable form.

target

The target image. Must have the same dimensionality as the source image.

targetMask

An optional mask image in target space, whose nonzero region will be the area over which the measure is calculated.

interpolation

A single integer specifying the type of interpolation to be applied to the source image when resampling it into the space of the target image. May be 0 (nearest neighbour), 1 (trilinear) or 3 (cubic spline). No other values are valid.

threads

For OpenMP-capable builds of the package, the maximum number of threads to use.

Value

A single numeric value representing the similarity between the images.

See also

Author

Jon Clayden <code@clayden.org>